Move subscriptions table to separate component
